+++ testutil/sample/crosschain.go
@@ -265,7 +265,6 @@ func GetCctxIndexFromString(index string) string {
 
 func CrossChainTx(t *testing.T, index string) *types.CrossChainTx {
 	r := newRandFromStringSeed(t, index)
-
 	return &types.CrossChainTx{
 		Creator:                 AccAddress(),
 		Index:                   GetCctxIndexFromString(index),
@@ -371,6 +370,24 @@ func InboundVote(coinType coin.CoinType, from, to int64) types.MsgVoteInbound {
 		TxOrigin:    EthAddress().String(),
 		Asset:       "",
 		EventIndex:  EventIndex(),
+	}
+}
+
+func OutboundVote(t *testing.T) types.MsgVoteOutbound {
+	cctx := CrossChainTx(t, EthAddress().String())
+	return types.MsgVoteOutbound{
+		CctxHash:                          cctx.Index,
+		OutboundTssNonce:                  cctx.GetCurrentOutboundParam().TssNonce,
+		OutboundChain:                     cctx.GetCurrentOutboundParam().ReceiverChainId,
+		Status:                            chains.ReceiveStatus_success,
+		Creator:                           cctx.Creator,
+		ObservedOutboundHash:              common.BytesToHash(EthAddress().Bytes()).String(),
+		ValueReceived:                     cctx.GetCurrentOutboundParam().Amount,
+		ObservedOutboundBlockHeight:       cctx.GetCurrentOutboundParam().ObservedExternalHeight,
+		ObservedOutboundEffectiveGasPrice: cctx.GetCurrentOutboundParam().EffectiveGasPrice,
+		ObservedOutboundGasUsed:           cctx.GetCurrentOutboundParam().GasUsed,
+		CoinType:                          cctx.InboundParams.CoinType,
+		ConfirmationMode:                  cctx.GetCurrentOutboundParam().ConfirmationMode,
 	}
 }
 
@@ -493,27 +510,6 @@ func CCTXfromRand(r *rand.Rand,
 		ProtocolContractVersion: ProtocolVersionFromRand(r),
 	}
 	return cctx
-}
-
-func OutboundVoteSim(r *rand.Rand,
-	cctx types.CrossChainTx,
-) (types.CrossChainTx, types.MsgVoteOutbound) {
-	msg := types.MsgVoteOutbound{
-		CctxHash:                          cctx.Index,
-		OutboundTssNonce:                  cctx.GetCurrentOutboundParam().TssNonce,
-		OutboundChain:                     cctx.GetCurrentOutboundParam().ReceiverChainId,
-		Status:                            chains.ReceiveStatus_success,
-		Creator:                           cctx.Creator,
-		ObservedOutboundHash:              common.BytesToHash(EthAddressFromRand(r).Bytes()).String(),
-		ValueReceived:                     cctx.GetCurrentOutboundParam().Amount,
-		ObservedOutboundBlockHeight:       cctx.GetCurrentOutboundParam().ObservedExternalHeight,
-		ObservedOutboundEffectiveGasPrice: cctx.GetCurrentOutboundParam().EffectiveGasPrice,
-		ObservedOutboundGasUsed:           cctx.GetCurrentOutboundParam().GasUsed,
-		CoinType:                          cctx.InboundParams.CoinType,
-		ConfirmationMode:                  cctx.GetCurrentOutboundParam().ConfirmationMode,
-	}
-
-	return cctx, msg
 }
 
 func ZRC20Withdrawal(to []byte, value *big.Int) *zrc20.ZRC20Withdrawal {